Add stylelint.utils
