+++ src/__tests__/plugins-test.js
@@ -2,19 +2,34 @@ import postcss from "postcss"
 import test from "tape"
 import stylelint from ".."
 
+const warnAboutFooMessages = stylelint.utils.ruleMessages("warn-about-foo", {
+  found: "found .foo",
+  notFound: "never found .foo",
+})
+
 function warnAboutFoo(expectation) {
   return (root, result) => {
     let foundFoo
     root.eachRule(rule => {
       if (rule.selector === ".foo") {
         if (expectation === "always") {
-          result.warn("found .foo")
+          stylelint.utils.report({
+            result,
+            ruleName: "warn-about-foo",
+            message: warnAboutFooMessages.found,
+            node: rule,
+          })
           foundFoo = true
         }
       }
     })
     if (!foundFoo) {
-      result.warn("never found .foo")
+      stylelint.utils.report({
+        result,
+        line: 1,
+        ruleName: "warn-about-foo",
+        message: warnAboutFooMessages.notFound,
+      })
     }
   }
 }
@@ -38,21 +53,25 @@ const configA = {
 }
 
 test("plugin runs", t => {
-  t.plan(4)
+  t.plan(6)
 
   const processor = postcss().use(stylelint(configA))
 
   processor.process(cssA)
     .then(result => {
       t.equal(result.warnings().length, 2)
-      t.equal(result.warnings()[0].text, "found .foo")
+      t.equal(result.warnings()[0].text, "found .foo (warn-about-foo)")
+      t.ok(result.warnings()[0].node)
     })
+    .catch(e => console.log(e.stack))
 
   processor.process(cssB)
     .then(result => {
       t.equal(result.warnings().length, 2)
-      t.equal(result.warnings()[0].text, "never found .foo")
+      t.equal(result.warnings()[0].text, "never found .foo (warn-about-foo)")
+      t.notOk(result.warnings()[0].node)
     })
+    .catch(e => console.log(e.stack))
 })
 
 const configB = {
@@ -75,6 +94,6 @@ test("plugin within set runs", t => {
   processor.process(cssA)
     .then(result => {
       t.equal(result.warnings().length, 2)
-      t.equal(result.warnings()[0].text, "found .foo")
+      t.equal(result.warnings()[0].text, "found .foo (warn-about-foo)")
     })
 })
